+++ frontend/src/utils/auth-client.ts
@@ -1,6 +1,4 @@
 import {
-  URI_ROOM_SVC,
-  URL_MATCHING_SVC,
   URL_USER_SVC,
   USER_CHANGE_PASSWORD,
   USER_CHANGE_USERNAME,
@@ -12,7 +10,6 @@ import {
 } from '../configs';
 import { requests, API } from './api-request';
 import { getTokens } from '../context/UserContext';
-import { QuestionModel } from '../components/room/QuestionModel';
 
 export const AuthClient = {
   signUp: (body: {
@@ -84,33 +81,5 @@ export const AuthClient = {
     username: string;
   }): Promise<API.Response<{ message: string }>> => {
     return requests.delete(URL_USER_SVC, USER_DELETE_USER, { data });
-  },
-
-  authRoom: (data: {
-    username: string;
-    room: string;
-    token: string;
-  }): Promise<API.Response<{ message: string }>> => {
-    const { username, room, token } = data;
-    const body = { username, roomId: room };
-    const headers = {
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
-      },
-    };
-
-    return requests.postWithHeaders(
-      URL_MATCHING_SVC,
-      '/api/matching/verify-user',
-      body,
-      headers
-    );
-  },
-
-  getQuestion: (data: {
-    room: string;
-  }): Promise<API.Response<{ question: QuestionModel }>> => {
-    return requests.get(URI_ROOM_SVC, `?roomId=${data.room}`);
   },
 };
